test: 인수테스트관련 공통 기능들을 추상화 클래스로 분리
