+++ src/test/java/nextstep/subway/acceptance/BaseAcceptanceTest.java
@@ -1,0 +1,24 @@
+package nextstep.subway.acceptance;
+
+import io.restassured.RestAssured;
+import nextstep.subway.acceptance.util.DatabaseCleanup;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public abstract class BaseAcceptanceTest {
+
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    private DatabaseCleanup databaseCleanup;
+
+    @BeforeEach
+    void setup() {
+        RestAssured.port = port;
+        databaseCleanup.execute();
+    }
+}
