+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -3,13 +3,9 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.acceptance.util.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
@@ -20,20 +16,10 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class LineAcceptanceTest {
-
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    private DatabaseCleanup databaseCleanup;
+public class LineAcceptanceTest extends BaseAcceptanceTest {
 
     @BeforeEach
-    void setup() {
-        RestAssured.port = port;
-        databaseCleanup.execute();
-
+    void dataSetUp() {
         StationAcceptanceTest.지하철역_등록_요청("강남역");
         StationAcceptanceTest.지하철역_등록_요청("역삼역");
         StationAcceptanceTest.지하철역_등록_요청("판교역");
